+++ src/ExponentialFamily.jl
@@ -52,6 +52,6 @@ include("distributions/laplace.jl")
 include("distributions/poisson.jl")
 include("distributions/chi_squared.jl")
 # include("distributions/mv_normal_wishart.jl")
-# include("distributions/normal_gamma.jl")
+include("distributions/normal_gamma.jl")
 
 end
